Professor Ratings Refactor
