+++ server/src/helpers/assistants/professorRatings/professorRatingsSchema.ts
@@ -1,41 +1,3 @@
-/**
- * A Zod schema that matches:
- *
- * {
- *   "name": "professor_ratings_schema",
- *   "schema": {
- *     "type": "object",
- *     "properties": {
- *       "results": {
- *         "type": "array",
- *         "items": {
- *           "type": "object",
- *           "properties": {
- *             "type": { "type": "string", "enum": [...] },
- *             "courses": { "type": ["array", "null"], "items": { "type": "string" } },
- *             "professors": { "type": ["array", "null"], "items": { "type": "string" } },
- *             "reason": { "type": ["string", "null"] }
- *           },
- *           "required": ["type", "courses", "professors", "reason"],
- *           "additionalProperties": false
- *         }
- *       }
- *     },
- *     "required": ["results"],
- *     "additionalProperties": false
- *   }
- * }
- */
-// export const ProfessorRatingsSchema = z.array(
-//   z
-//     .object({
-//       type: z.enum(["professor", "courses", "both", "fallback"]),
-//       courses: z.array(z.string()).nullable(),
-//       professors: z.array(z.string()).nullable(),
-//       reason: z.string().nullable(),
-//     })
-//     .strict()
-// );
 export const ProfessorRatingsSchema = {
   type: "object",
   properties: {
@@ -61,27 +23,11 @@ export const ProfessorRatingsSchema = {
   additionalProperties: false,
 };
 
-/*
-example:
-    schema: {
-        type: "object",
-        properties: {
-          name: { 
-            type: "string" 
-          },
-          date: { 
-            type: "string" 
-          },
-          participants: { 
-            type: "array", 
-            items: { 
-              type: "string" 
-            } 
-          },
-        },
-        required: ["name", "date", "participants"],
-        additionalProperties: false,
-      },
-    }
-  }
-*/
+export type ProfessorRatingsResponse = {
+  results: {
+    type: "professor" | "courses" | "both" | "fallback";
+    courses: string[] | null;
+    professors: string[] | null;
+    reason: string | null;
+  }[];
+} | null;
