+++ shared/src/types/assistant/index.ts
@@ -4,6 +4,7 @@ export type AssistantType = {
   id: string; // The unique id of the assistant
   title: string; // the title of the assistant
   instructions?: string; // the instructions of the assistant
+  helperInstructions?: string; // the helper instructions of the assistant
   assistantId?: string; // The unique id of the assistant
   desc: string; // the displayed description of the assistant
   urlPhoto?: string; // The avatar image of the assistant
